web accessibility id

+++ ohos_nweb/include/nweb.h
@@ -352,7 +352,7 @@ public:
      *
      * @param accessibilityIdGenerator Accessibility id generator.
      */
-    virtual void PutAccessibilityIdGenerator(std::function<int32_t()> accessibilityIdGenerator) = 0;
+    virtual void PutAccessibilityIdGenerator(std::function<int64_t()> accessibilityIdGenerator) = 0;
 
     /**
      * Set the NWebHandler that will receive various notifications and
@@ -876,7 +876,7 @@ public:
      * @param accessibilityId The id of the accessibility node.
      * @param action The action to be performed on the accessibility node.
      */
-    virtual void ExecuteAction(int32_t accessibilityId, uint32_t action) const = 0;
+    virtual void ExecuteAction(int64_t accessibilityId, uint32_t action) const = 0;
 
     /**
      * Get the information of the focused accessibility node on the given accessibility node in the browser.
@@ -887,7 +887,7 @@ public:
      * @return true if get accessibility node info successfully, otherwise false.
      */
     virtual bool GetFocusedAccessibilityNodeInfo(
-        int32_t accessibilityId, bool isAccessibilityFocus, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
+        int64_t accessibilityId, bool isAccessibilityFocus, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
 
     /**
      * Get the information of the accessibility node by its accessibility id in the browser.
@@ -896,7 +896,7 @@ public:
      * @return true if get accessibility node info successfully, otherwise false.
      */
     virtual bool GetAccessibilityNodeInfoById(
-        int32_t accessibilityId, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
+        int64_t accessibilityId, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
 
     /**
      * Get the information of the accessibility node by focus move in the browser.
@@ -906,7 +906,7 @@ public:
      * @return true if get accessibility node info successfully, otherwise false.
      */
     virtual bool GetAccessibilityNodeInfoByFocusMove(
-        int32_t accessibilityId, int32_t direction, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
+        int64_t accessibilityId, int32_t direction, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
 
     /**
      * Set the accessibility state in the browser.
